more simple use of AdboudDialog to get modules ... for us
